+++ src/Item/Bow.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "Item/Item.hpp"
 #include "Core/Math.hpp"
+#include "Item/Item.hpp"
 
 class BowItem : public Item
 {
@@ -9,6 +9,6 @@ class BowItem : public Item
 
 public:
     BowItem();
-    virtual void interact(World& world, size_t dimension, ItemStack& stack, glm::i64vec3 pos, glm::i64vec3 normal, InventoryContainer& inventory) override;
-    virtual void on_release(World& world, size_t dimension, ItemStack& stack, glm::i64vec3 pos, glm::vec3 dir, InventoryContainer& inventory) override;
+    virtual void interact(World& world, int dimension, ItemStack& stack, glm::i64vec3 pos, glm::i64vec3 normal, InventoryContainer& inventory) override;
+    virtual void on_release(World& world, int dimension, ItemStack& stack, glm::i64vec3 pos, glm::vec3 dir, InventoryContainer& inventory) override;
 };
