+++ src/Item/Bucket.cpp
@@ -9,7 +9,7 @@ BucketItem::BucketItem()
     set_texture(Engine::get().registry().create_texture("assets/textures/water.png"));
 }
 
-void BucketItem::interact(World& world, size_t dimension, ItemStack& stack, glm::i64vec3 pos, glm::i64vec3 normal, InventoryContainer& inventory)
+void BucketItem::interact(World& world, int dimension, ItemStack& stack, glm::i64vec3 pos, glm::i64vec3 normal, InventoryContainer& inventory)
 {
     (void)stack;
     (void)inventory;
